transform: fix incorrect alignment of heap-to-stack transform
It assumed the maximum alignment was equal to sizeof(void*), which is
definitely not the case. So this only worked more or less by accident
previously.

I found this bug while working on MIPS support, but it is also a bug on
many other architectures like amd64 (which has a maximum alignment of 16
bytes, not 8 as would be assumed here).