compiler: use memcmp to compare strings
This replaces our runtime stringEqual and stringLess functions with calls to libc's memcmp.
This has a few advantages:
1. Memory comparison functions are not duplicated between Go and C
2. LLVM can optimize the compare-to-empty case by itself, so OptimizeStringEqual is no longer needed
3. LLVM can rewrite small constant-length memcmp operations into direct loads + compares
4. Compares to constants are generally compiled into simpler IR

The downside of this is that all of the comparison logic must be handled by the compiler frontend.
For string equality checks, we only need to compare the lengths and memcmp if they are equal.
String order checks are messier, since we need to:
1. Find the minimum length
2. Call memcmp with the minimum length
3. Compare the lengths
4. Merge the length comparison with the memory comparison